requirements.txt
ursina
perlin-noise
numpy
pyinstaller
screeninfo
Pillow
